modify: sort DML values by key name
